backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
openpyxl==3.1.2
pandas==2.1.4
python-multipart==0.0.6
pydantic==2.5.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.5.3
requests==2.31.0
